requirements.txt
# Discord bot dependencies
discord.py>=2.0.0

# Google Gemini AI integration
google-generativeai>=0.3.0

# Environment variable management
python-dotenv>=1.0.0
